+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added Undo action for Snackbar on deleting Healing (and Payment)
 
 ## [2.0.0-alpha04] - 2020-02-21
 ### Changed